+++ Runtime/InventoryEditorPreset.cs
@@ -19,28 +19,82 @@ namespace OcDialogue
         static InventoryEditorPreset _instance;
 
         [InitializeOnLoadMethod]
-        static void Init()
+        static void EditorInit()
+        {
+            _instance = Resources.Load<InventoryEditorPreset>(AssetPath);
+        }
+
+        [RuntimeInitializeOnLoadMethod]
+        static void RuntimeInit()
         {
             _instance = Resources.Load<InventoryEditorPreset>(AssetPath);
         }
 
         public bool usePreset;
         [TableList]public List<ItemPreset> ItemPresets;
+        
+        [PreviewField(200f, ObjectFieldAlignment.Left), OnValueChanged("OnItemDropped"), LabelWidth(80)] public UnityEngine.Object dropDesk;
+        [OnCollectionChanged("OnArrayDropped"), LabelWidth(80)] public List<UnityEngine.Object> arrayDropDesk;
+        
+        void OnItemDropped(UnityEngine.Object target)
+        {
+            var item = target as ItemBase;
+            if (item == null) return;
+            var preset = new ItemPreset()
+            {
+                item = item,
+                icon = item.editorIcon,
+                type = item.type,
+                
+            };
+            switch (item.type)
+            {
+                case ItemType.Generic:
+                    preset.genericType = (GenericType) Enum.Parse(typeof(GenericType), item.SubTypeString);
+                    break;
+                case ItemType.Armor:
+                    preset.armorType = (ArmorType) Enum.Parse(typeof(ArmorType), item.SubTypeString);
+                    break;
+                case ItemType.Weapon:
+                    preset.weaponType = (WeaponType) Enum.Parse(typeof(WeaponType), item.SubTypeString);
+                    break;
+                case ItemType.Accessory:
+                    preset.accessoryType = (AccessoryType) Enum.Parse(typeof(AccessoryType), item.SubTypeString);
+                    break;
+                case ItemType.Important:
+                    preset.importantItemType = (ImportantItemType) Enum.Parse(typeof(ImportantItemType), item.SubTypeString);
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+            ItemPresets.Add(preset);
+            dropDesk = null;
+        }
+
+        void OnArrayDropped()
+        {
+            OnItemDropped(arrayDropDesk[0]);
+            arrayDropDesk.Clear();
+        }
+
+
         [Serializable]
         public class ItemPreset
         {
-            [TableColumnWidth(48, false), PreviewField] public UnityEngine.Object icon;
-            [VerticalGroup("Type"), TableColumnWidth(200, false)]public ItemType type;
+            [TableColumnWidth(48, false), PreviewField, OnValueChanged("OnItemDropped")] public UnityEngine.Object icon;
+            [VerticalGroup("Type"), TableColumnWidth(200, false), OnValueChanged("OnTypeChanged")]public ItemType type;
             [HideInInspector] public string subtypeString;
-            [VerticalGroup("Type"), OnValueChanged("ApplySubTypeString"), HideLabel][ShowIf("type", ItemType.Generic)]public GenericType genericType;
-            [VerticalGroup("Type"), OnValueChanged("ApplySubTypeString"), HideLabel][ShowIf("type", ItemType.Armor)]public ArmorType armorType;
-            [VerticalGroup("Type"), OnValueChanged("ApplySubTypeString"), HideLabel][ShowIf("type", ItemType.Weapon)]public WeaponType weaponType;
-            [VerticalGroup("Type"), OnValueChanged("ApplySubTypeString"), HideLabel][ShowIf("type", ItemType.Important)]public ImportantItemType importantItemType;
+            [VerticalGroup("Type"), OnValueChanged("OnTypeChanged"), HideLabel][ShowIf("type", ItemType.Generic)]public GenericType genericType;
+            [VerticalGroup("Type"), OnValueChanged("OnTypeChanged"), HideLabel][ShowIf("type", ItemType.Armor)]public ArmorType armorType;
+            [VerticalGroup("Type"), OnValueChanged("OnTypeChanged"), HideLabel][ShowIf("type", ItemType.Weapon)]public WeaponType weaponType;
+            [VerticalGroup("Type"), OnValueChanged("OnTypeChanged"), HideLabel][ShowIf("type", ItemType.Important)]public ImportantItemType importantItemType;
+            [VerticalGroup("Type"), OnValueChanged("OnTypeChanged"), HideLabel][ShowIf("type", ItemType.Accessory)]public AccessoryType accessoryType;
             
             [VerticalGroup("Item")][ValueDropdown("GetItemList"), OnValueChanged("ApplyIcon"), HideLabel]public ItemBase item;
             [VerticalGroup("Item")][ShowIf("@item != null && item.isStackable")] public int count;
 
-            void ApplySubTypeString()
+            [HideInInspector] public UnityEngine.Object _cachedIcon;
+            void OnTypeChanged()
             {
                 switch (type)
                 {
@@ -64,6 +118,40 @@ namespace OcDialogue
             void ApplyIcon()
             {
                 icon = item.editorIcon;
+                _cachedIcon = icon;
+            }
+
+            void OnItemDropped()
+            {
+                if (_cachedIcon != icon)
+                {
+                    item = icon as ItemBase;
+                }
+                
+                
+                icon = item.editorIcon;
+
+                type = item.type;
+                switch (type)
+                {
+                    case ItemType.Generic:
+                        genericType = (GenericType) Enum.Parse(typeof(GenericType), item.SubTypeString);
+                        break;
+                    case ItemType.Armor:
+                        armorType = (ArmorType) Enum.Parse(typeof(ArmorType), item.SubTypeString);
+                        break;
+                    case ItemType.Weapon:
+                        weaponType = (WeaponType) Enum.Parse(typeof(WeaponType), item.SubTypeString);
+                        break;
+                    case ItemType.Accessory:
+                        accessoryType = (AccessoryType) Enum.Parse(typeof(AccessoryType), item.SubTypeString);
+                        break;
+                    case ItemType.Important:
+                        importantItemType = (ImportantItemType) Enum.Parse(typeof(ImportantItemType), item.SubTypeString);
+                        break;
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
             }
 
             ValueDropdownList<ItemBase> GetItemList()
@@ -76,6 +164,14 @@ namespace OcDialogue
                 }
 
                 return vdList;
+            }
+
+            public ItemBase GetCopy()
+            {
+                var copy = item.GetCopy();
+                if (copy.isStackable) copy.AddStack(count);
+
+                return copy;
             }
         }
 #endif